# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Kaay entreprendre",
        "description": "Cet api va nous permettre de faire communiquer la partie front et la partie back de notre plateforme",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/ressources": {
            "get": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Obtenir la liste des ressources",
                "description": "Récupère la liste de toutes les ressources non supprimées.",
                "operationId": "getRessources",
                "responses": {
                    "200": {
                        "description": "Liste des ressources récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/ajouter-ressource": {
            "post": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Ajouter une nouvelle ressource",
                "description": "Ajoute une nouvelle ressource avec les détails fournis.",
                "operationId": "ajouterRessource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "lien": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ressource ajoutée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/ressources/{id}": {
            "post": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Modifier une ressource existante",
                "description": "Modifie les détails d'une ressource existante en fonction de l'ID fourni.",
                "operationId": "modifierRessource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titre": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "lien": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ressource modifiée avec succès"
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Supprimer une ressource",
                "description": "Supprime une ressource en fonction de l'ID fourni.",
                "operationId": "supprimerRessource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ressource supprimée avec succès"
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/ressource": {
            "get": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Afficher une ressource",
                "description": "Affiche les détails d'une ressource en fonction de l'ID fourni.",
                "operationId": "getRessource",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID de la ressource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la ressource récupérés avec succès"
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/ressource/archive": {
            "post": {
                "tags": [
                    "Ressources"
                ],
                "summary": "Archiver une ressource",
                "description": "Archive une ressource en fonction de l'ID fourni.",
                "operationId": "archiverRessource",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ressource archivée avec succès"
                    },
                    "404": {
                        "description": "Ressource non trouvée"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}